Throwaway garbage code before ditching branch
Normal maps turn out to be a bit too complicated for what they are
bringing to the table *right now*.
In particular, it would add quite a bit of complexity to handle mesh
resource reuse at a time where the whole mesh direction is not yet so
clear.
Using the THREE.js-provided normal and displacement map might not work
so well for this, maybe just uploading a texture whose alpha is
displacement would work better, or even as a uniform (might be a
garbage idea).
There is also the problem of the inability of trivially encoding the
normals with 8-bit integers because of the Z-heavy values, which could
probably be solved but I can't be arsed right now.